update and create products endpoints

+++ src/main/java/dev/abhishek/ecommerce/modules/product/dto/UpdateProductRequest.java
@@ -28,5 +28,5 @@ public class UpdateProductRequest {
     @NotNull
     private String description;
 
-    private Integer category_id;
+    private Long category_id;
 }
